tests/integration/apps/cassandra-app/requirements.txt
starlette
uvicorn
cassandra-driver
